+++ src/tools/post-build.ts
@@ -52,17 +52,17 @@ const repo = "Sv443/BetterYTM";
 const userscriptDistFile = `BetterYTM${suffix}.user.js`;
 const distFolderPath = `./${rollupCfgOutputDir}/`;
 const assetFolderPath = "./assets/";
-const scriptUrl = (() => {
-  switch(host) {
-  case "greasyfork":
-    return "https://update.greasyfork.org/scripts/475682/BetterYTM.user.js";
-  case "openuserjs":
-    return "https://openuserjs.org/install/Sv443/BetterYTM.user.js";
-  case "github":
-  default:
-    return `https://raw.githubusercontent.com/${repo}/main/dist/${userscriptDistFile}`;
-  }
-})();
+// const hostScriptUrl = (() => {
+//   switch(host) {
+//   case "greasyfork":
+//     return "https://update.greasyfork.org/scripts/475682/BetterYTM.user.js";
+//   case "openuserjs":
+//     return "https://openuserjs.org/install/Sv443/BetterYTM.user.js";
+//   case "github":
+//   default:
+//     return `https://raw.githubusercontent.com/${repo}/main/dist/${userscriptDistFile}`;
+//   }
+// })();
 
 /** Whether to trigger the bell sound in some terminals when the code has finished compiling */
 const ringBell = Boolean(env.RING_BELL && (env.RING_BELL.length > 0 && env.RING_BELL.trim().toLowerCase() === "true"));
@@ -96,8 +96,6 @@ ${localizedDescriptions ? "\n" + localizedDescriptions : ""}\
 // @match             https://music.youtube.com/*
 // @match             https://www.youtube.com/*
 // @run-at            document-start
-// @downloadURL       ${scriptUrl}
-// @updateURL         ${scriptUrl}
 // @connect           api.sv443.net
 // @connect           github.com
 // @connect           raw.githubusercontent.com