+++ changelog.md
@@ -35,6 +35,7 @@
       - From `watchMetadata` to `ytWatchMetadata`
   - Added Storybook for easier and faster development of components
   - Improved ripple component design and functionality (refer to the contributing guide for more info)
+  - Removed the `@updateURL` and `@downloadURL` directives because their use is controversial and the script has a built-in update check now
 - **Plugin Interface Changes:**
   - Added new components:
     -  `createLongBtn()` to create a button with an icon and text (works either as normal or as a toggle button)